Separate the implementation of initializeRecaptchaConfig (#7515)
