+++ packages/auth/src/core/auth/auth_impl.test.ts
@@ -40,6 +40,7 @@ import * as navigator from '../util/navigator';
 import * as reload from '../user/reload';
 import { AuthImpl, DefaultConfig } from './auth_impl';
 import { _initializeAuthInstance } from './initialize';
+import { _initializeRecaptchaConfig } from '../../platform_browser/recaptcha/recaptcha_enterprise_verifier';
 import { ClientPlatform } from '../util/version';
 import { mockEndpointWithParams } from '../../../test/helpers/api/helper';
 import { Endpoint, RecaptchaClientType, RecaptchaVersion } from '../../api';
@@ -734,7 +735,7 @@ describe('core/auth/auth_impl', () => {
         },
         recaptchaConfigResponseEnforce
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       expect(auth._getRecaptchaConfig()).to.eql(cachedRecaptchaConfigEnforce);
     });
@@ -751,7 +752,7 @@ describe('core/auth/auth_impl', () => {
         },
         recaptchaConfigResponseOff
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       expect(auth._getRecaptchaConfig()).to.eql(cachedRecaptchaConfigOFF);
     });
@@ -767,7 +768,7 @@ describe('core/auth/auth_impl', () => {
         },
         recaptchaConfigResponseEnforce
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
       auth.tenantId = 'tenant-id';
       mockEndpointWithParams(
         Endpoint.GET_RECAPTCHA_CONFIG,
@@ -778,7 +779,7 @@ describe('core/auth/auth_impl', () => {
         },
         recaptchaConfigResponseOff
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       auth.tenantId = null;
       expect(auth._getRecaptchaConfig()).to.eql(cachedRecaptchaConfigEnforce);
