+++ packages/auth/src/core/strategies/email_and_password.test.ts
@@ -50,6 +50,7 @@ import {
   verifyPasswordResetCode
 } from './email_and_password';
 import { MockGreCAPTCHATopLevel } from '../../platform_browser/recaptcha/recaptcha_mock';
+import { _initializeRecaptchaConfig } from '../../platform_browser/recaptcha/recaptcha_enterprise_verifier';
 
 use(chaiAsPromised);
 use(sinonChai);
@@ -202,7 +203,7 @@ describe('core/strategies/sendPasswordResetEmail', () => {
         },
         recaptchaConfigResponseEnforce
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       const apiMock = mockEndpoint(Endpoint.SEND_OOB_CODE, {
         email
@@ -230,7 +231,7 @@ describe('core/strategies/sendPasswordResetEmail', () => {
         },
         recaptchaConfigResponseOff
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       const apiMock = mockEndpoint(Endpoint.SEND_OOB_CODE, {
         email
@@ -299,7 +300,7 @@ describe('core/strategies/sendPasswordResetEmail', () => {
         },
         recaptchaConfigResponseEnforce
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       mockEndpoint(Endpoint.SEND_OOB_CODE, { email });
       const response = await sendPasswordResetEmail(auth, email);
@@ -617,7 +618,7 @@ describe('core/strategies/email_and_password/createUserWithEmailAndPassword', ()
         },
         recaptchaConfigResponseEnforce
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       const { _tokenResponse, user, operationType } =
         (await createUserWithEmailAndPassword(
@@ -648,7 +649,7 @@ describe('core/strategies/email_and_password/createUserWithEmailAndPassword', ()
         },
         recaptchaConfigResponseEnforce
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       const { _tokenResponse, user, operationType } =
         (await createUserWithEmailAndPassword(
@@ -726,7 +727,7 @@ describe('core/strategies/email_and_password/createUserWithEmailAndPassword', ()
         },
         recaptchaConfigResponseEnforce
       );
-      await auth.initializeRecaptchaConfig();
+      await _initializeRecaptchaConfig(auth);
 
       const { _tokenResponse, user, operationType } =
         (await createUserWithEmailAndPassword(
