+++ packages/auth/src/model/auth.ts
@@ -97,5 +97,4 @@ export interface AuthInternal extends Auth {
 
   useDeviceLanguage(): void;
   signOut(): Promise<void>;
-  initializeRecaptchaConfig(): Promise<void>;
 }
